Enable multiple messages in hardware fifo.
This change enables higher performance and stability especially for high throughput connections. This is archived by switching from allowing only a single message in the hardware fifo (because it is flushed completly after each message), to multiple message in buffer, reading exactly as much bytes as necessary and leaving further bytes untouched. Note that it is assumed that the length byte is correct and as much bytes are expected to be received / read from the fifo. This should hold true because the hardware would read / fill as much bytes to the fifo as the length byte indicate, even if they are incorrect (which should be cought by the crc afterwards).